[#2526] Move actions into separate files
