+++ amy/emails/actions/__init__.py
@@ -1,0 +1,18 @@
+# flake8: noqa
+from emails.actions.admin_signs_instructor_up_for_workshop import (
+    admin_signs_instructor_up_for_workshop_receiver,
+)
+from emails.actions.instructor_badge_awarded import instructor_badge_awarded_receiver
+from emails.actions.instructor_confirmed_for_workshop import (
+    instructor_confirmed_for_workshop_receiver,
+)
+from emails.actions.instructor_declined_from_workshop import (
+    instructor_declined_from_workshop_receiver,
+)
+from emails.actions.instructor_signs_up_for_workshop import (
+    instructor_signs_up_for_workshop_receiver,
+)
+from emails.actions.instructor_training_approaching import (
+    instructor_training_approaching_receiver,
+)
+from emails.actions.persons_merged import persons_merged_receiver
